+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.0
+
+- [refactor] Using ICE PKG to build the package.
+- [feature] Support minimum version of appear-polyfill.
+- [refactor] Support passive event listener.
 ## 0.1.3
 
 - [feat] Support custom option `endReachedThreshold`